Fix: change float to double

+++ headers/ft_dda.h
@@ -6,7 +6,7 @@
 /*   By: tlassere <tlassere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 13:23:52 by tlassere          #+#    #+#             */
-/*   Updated: 2024/05/02 15:27:51 by tlassere         ###   ########.fr       */
+/*   Updated: 2024/05/02 15:35:03 by tlassere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,9 +38,9 @@ typedef struct s_collide_data
 	t_ivec	block_cheked;
 }	t_collide_data;
 
-t_collide_data	ft_dda_call(t_data *data, t_dvec ray_start, float rotat,
+t_collide_data	ft_dda_call(t_data *data, t_dvec ray_start, double rotat,
 					int (*check_block)(int));
-t_collide_data	ft_dda(t_data *data, t_dvec ray_start, float rotat);
-t_collide_data	ft_dda_interact(t_data *data, t_dvec ray_start, float rotat);
+t_collide_data	ft_dda(t_data *data, t_dvec ray_start, double rotat);
+t_collide_data	ft_dda_interact(t_data *data, t_dvec ray_start, double rotat);
 
 #endif
