+++ sources/ft_game/ft_dda_call.c
@@ -6,7 +6,7 @@
 /*   By: tlassere <tlassere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 13:22:20 by tlassere          #+#    #+#             */
-/*   Updated: 2024/05/02 15:27:51 by tlassere         ###   ########.fr       */
+/*   Updated: 2024/05/02 15:34:39 by tlassere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -114,7 +114,7 @@ static t_collide_data	ft_collide(t_ray_data ray, t_data *data,
 }
 
 t_collide_data	ft_dda_call(t_data *data, t_dvec ray_start,
-	float rotat, int (*check_block)(int))
+	double rotat, int (*check_block)(int))
 {
 	double			rotat_op;
 	t_ray_data		ray;
